Capstone_Project/src/test/java/Capstone_Project_Test/EnergyManagementTest.java
package Capstone_Project_Test;

import static org.mockito.Mockito.*;
import static org.junit.Assert.*;
import org.junit.Before;
import org.junit.Test;

import Capstone_Project.ChargingStationGUI;
import Capstone_Project.EnergyManagement;
import Capstone_Project.Weather;

public class EnergyManagementTest {

    private EnergyManagement energyManagement;
    private Weather mockWeather;
    private ChargingStationGUI mockGui;

    @Before
    public void setUp() {
        mockWeather = mock(Weather.class);
        mockGui = mock(ChargingStationGUI.class);
        energyManagement = new EnergyManagement(mockWeather, mockGui);
    }

    @Test
    public void testEnergySourceIsSolarWhenSunny() {
        when(mockWeather.isSunny()).thenReturn(true);
        String energySource = energyManagement.getCurrentEnergySource();
        assertEquals("Energy source should be Solar when it's sunny", "Solar", energySource);
    }

    @Test
    public void testEnergySourceIsGridWhenNotSunny() {
        when(mockWeather.isSunny()).thenReturn(false);
        String energySource = energyManagement.getCurrentEnergySource();
        assertEquals("Energy source should be Grid when it's not sunny", "Grid", energySource);
        verify(mockGui).updateEnergySource("Grid");
    }
}
